Clear the findings from the first analyze pass
The pass reported 69 unique findings, 181 across translation units.
Most of them collapse to a handful of causes.

C6297 accounted for 108 and had a single origin. In INDEX_BY_VALUE64
and index_by_value, "(u32)((K << 1) + ((VAL >> ...) & 1))" adds a u32
shift to a 64-bit value, so the shift is widened to 64 bits, which is
what the warning names. Narrowing the second term instead keeps the
addition in 32 bits and leaves the result unchanged. The 28 call sites
in arena.c, arena_dbg.c and the suites needed no edit.

C6340 was 25 uses of %d for a u32 in test format strings. One of them,
in arena_test.c, also passed arguments in an order its own labels did
not match: the log2 exponent printed under "Expected" and the expected
index under "Actual". C6001 was a local assigned only inside FL_TRY and
FL_CATCH bodies and read after them, in arena_malloc.c and in four
fault_injector_tests.c cases. A catch is reachable before the
assignment, so each local now has an initializer. The unchecked _dup2
in fl_log_service_tests.c is now checked.

The two ill-defined-loop reports in chunk_tests.c come from KIBI, which
yields long long: the mixed comparison makes the analyzer mis-model a
size_t counter, and casting the bound restores it.

Seven findings are wrong about code that is already correct. Each now
carries FL_ANALYSIS_SUPPRESS, a new macro in fl_macros.h that expands
to __pragma(warning(suppress)) on MSVC and to nothing elsewhere, plus a
comment saying what rules the finding out: a decommit inside a region
that is released later, a write count that is exactly argc, a pointer
the throw above it makes unreachable, a loop that relies on u16
wraparound, a 64 KB buffer that stays on the stack because malloc in
that unit is the injectable memory service rather than the CRT's, and
two declarations that must stay unannotated in order to replace the
CRT's.

+++ src/chunk_tests.c
@@ -480,7 +480,9 @@ FL_TEST("Maximum Request", max_request) {
  * should be a great enough range to ensure this macro is correct.
  */
 FL_TEST("Request to Size", request_to_size) {
-    for (size_t req = 0; req <= KIBI(8); req++) {
+    // (size_t): KIBI yields long long, and the mixed comparison makes /analyze
+    // mis-model the loop bound.
+    for (size_t req = 0; req <= (size_t)KIBI(8); req++) {
         request_to_size_is_valid(req);
     }
 }
@@ -515,7 +517,9 @@ FL_TEST("Pad Request", pad_request) {
  * less the size of a ChunkHeader.
  */
 FL_TEST("Payload vs Chunk Size", payload_vs_chunk_size) {
-    for (size_t req = 0; req < KIBI(64); req++) {
+    // (size_t): KIBI yields long long, and the mixed comparison makes /analyze
+    // mis-model the loop bound.
+    for (size_t req = 0; req < (size_t)KIBI(64); req++) {
         size_t chunk_size = CHUNK_SIZE_FROM_REQUEST(req);
         size_t expected   = chunk_size - CHUNK_ALIGNED_SIZE;
         size_t actual     = CHUNK_REQUEST_TO_PAYLOAD(req);
